Rename database module to btree and implement btree iterator
This renames the database module to btree to increase the similarity
between our module and the original sqlite implementation.

Besides the database module implementation, this also renames the
SchemaCell struct in the record module to Schema, and the CellInfo
struct to Cell. It's also created a struct called Entry for representing the leaf
data of Cell.

Entry comes with a iterator, who iterates over the entries of the
b-tree, i.e. the data of the leaf cells.